fix(types): widen PaginatedList.total to i64 with #[ts(type = "number")] override
Reverts the i32 narrowing from the initial Codex fix. Keeping the Rust
field wide (i64) avoids baking a 2B cap into a generic pagination type.
The #[ts(type = "number")] attribute overrides ts-rs to emit `number`
instead of `bigint` in the generated TypeScript binding.

+++ crates/types/src/pagination.rs
@@ -5,14 +5,15 @@ use ts_rs::TS;
 #[ts(export)]
 pub struct PaginatedList<T: TS> {
     pub items: Vec<T>,
-    pub total: i32,
+    #[ts(type = "number")]
+    pub total: i64,
     pub page: u32,
     pub per_page: u32,
     pub total_pages: u32,
 }
 
 impl<T: TS> PaginatedList<T> {
-    pub fn new(items: Vec<T>, total: i32, page: u32, per_page: u32) -> Self {
+    pub fn new(items: Vec<T>, total: i64, page: u32, per_page: u32) -> Self {
         let total_pages = if total <= 0 || per_page == 0 {
             0
         } else {
